Combined the promises into the youtube process.
