--- .jscsrc
@@ -1,29 +1,0 @@
-{
-    "maximumLineLength": 80,
-    "requireCurlyBraces": ["if", "else", "for", "while", "do", "try", "catch"],
-    "disallowSpaceAfterKeywords": ["if", "else", "for", "while", "do", "try", "catch"],
-    "requireSpacesInAnonymousFunctionExpression": {"beforeOpeningCurlyBrace": true},
-    "disallowSpacesInAnonymousFunctionExpression": {"beforeOpeningRoundBrace": true},
-    "requireSpacesInNamedFunctionExpression": {"beforeOpeningCurlyBrace": true},
-    "disallowSpacesInNamedFunctionExpression": {"beforeOpeningRoundBrace": true},
-    "requireSpacesInFunctionDeclaration": {"beforeOpeningCurlyBrace": true},
-    "disallowSpacesInFunctionDeclaration": {"beforeOpeningRoundBrace": true},
-    "disallowSpacesInCallExpression": true,
-    "requireBlocksOnNewline": true,
-    "disallowEmptyBlocks": true,
-    "disallowDanglingUnderscores": true,
-    "disallowSpaceAfterObjectKeys": true,
-    "requireSpaceBeforeObjectValues": true,
-    "requireCommaBeforeLineBreak": true,
-    "disallowSpaceAfterPrefixUnaryOperators": true,
-    "disallowSpaceBeforePostfixUnaryOperators": true,
-    "requireSpaceBeforeBinaryOperators": true,
-    "requireSpaceAfterBinaryOperators": true,
-    "requireCamelCaseOrUpperCaseIdentifiers": true,
-    "disallowMultipleLineBreaks": true,
-    "disallowTrailingWhitespace": true,
-    "disallowTrailingComma": true,
-    "requireCapitalizedConstructors": true,
-    "requireSpaceAfterLineComment": true,
-    "requireNewlineBeforeBlockStatements": true
-}
